Create Basic IoT
This project can be controlled using our smart phone or Laptop from anywhere in this world. This project uses MQTT protocol for data transfer. For this project, I have used Ubidots MQTT server as it gives access to make several variables for free. So I have created 16 variables on my Ubidots server. And I have programmed my ESP32 board as an MQTT client which will receive the data from MQTT broker and according to the data received, the relay will be switched and the respected appliances can be controlled easily.
As this is a cloud based MQTT server so we can easily control the Appliances from anywhere in this world. You can login to your Ubidots account from smartphone as well as Personal computer and can control/monitor the appliances on the go.